Implemented leader election
Got a lot of boilerplate code in place.
Leader election implemented as described in the Raft Paper.
But without any log, so nodes just vote for the highest term.
When a leader is elected it will keep sending AppendEntries RPCs till it sees a larger term.
Followers will not start a new election as long as they receive repeated AppendEntries RPCs.