RequestVoteResponse now includes request term
We have no way of knowing in which term the request was sent. We solve
this by including it in the response. The requester might have increased
its term before a response is received, this case is however handled by
the algorithm so it doesn't matter.
RequestVoteQF should now be able to return as soon as a higher term is
seen instead of needing a minimum of two replies. Since the local node
is no longer part of the configuration there might have been cases were
a higher term would not have been detected. We would have had to wait
for all nodes to respond before realizing that the vote failed.