Remove dependency on gorums/idutil
As idutil can give different ids for the same service, i.e., localhost
gives a different hash than 127.0.0.1, we switch to using user provided
ids. We now depend on the order in which the servers where specified,
i.e., a cluster with id 2 refers to the second server in the cluster
argument. This works for a fixed configuration. When we move on to
reconfiguration we can change this to id/name=address pairs, as there
will be gaps in the cluster slice when we remove servers.

Note: As we cannot get back the nodes created by the Manager in the
order they were given, we still need to use the hashing function to
create a mapping between the ids.